feat: ability to use matrix elements in decays
Matrix elements can be implemented for the phase space decays.
Previously, to achieve the uniform sampling the sampled are
rejected with respect of their phase space weights.
This is equivalent to having a matrix element equal to one.
Now a function pointer can be set which repesents a matrix element.
This function will then modify the weights with a more physical result.
The default functio is still a function returning always one.