refactor: set matrix element only in constructor.
Removed the SetMatrixElement method, because it's kind
of impossible to decide if its different from the default.
(can't get plain function pointer out of std::function).
But i need to know if its the default, because the estimate
method in the broad phase to determine min max values of the
phase space differs by using default or non default matrix elements.