[alias] finalize cpp11 developement
correct unittest-files generator-scripts
add missing continuous randomizer in pybindings
the propagator test will only propagate 10 muons instead of 100,
so the testfile zip, dominated in size by this single unit test,
is not that large
add new testfiles

the warning -Wformat-security arising at the definition of the logger macros
was only occuring in clang because there the warning is checked by default.
In gcc it's not in -Wall.
I now ignored this warning for clang, but the reason, why it arises is still unknown
and one should find it out in the future.