[alias] Remove sign of particle charge from Interpolation tables
Up to now, there is no process in the cross sections,
which differ between positiv and negativ charge of the propagating particle.
(weak interaction was merged after this fork was made)
The only part, where the charge was not squared, was in the multiple scattering.
But there it was multiplied to the gaussian function,
which is symmetric around zero,
which means, that it doesn't matter, which part is chosen in the monte carlo propagation.
This will of course break the UnitTest, because it is not the same propagation, as before,
but the physics is the same.
So new TestFiles have to be produced.
But there has to be some new ones before the next release anyway, because of the ClosestApproachPoint fix ;)