fix: separate secondary output and logging
the problem when calling PROPOSAL as an extern library were the macros set in the PROPOSAL cmake but not in extrenal cmakes.
The IceCube macro was needed for the Logging.
Now the Logging is not in the public includes and just used in the source files, not in the headers.