Avoid pulling random numbers when scattering when they are not needed.
Scattering will now return a pair of averaged scattering direction
and the direction of the particle at the end of propagation step.
In "AdvanceParticle" scattering (direction change) and displacement will
from now on consider separately.

The testfile must be modified, since scattering no longer takes over
the displacement of the particle.