Add UnitTests for UtilityIntegral and changes in IntegralWithLog function
When using IntegrateWithLog with an absolute randomRatio (e.g. randomRatio<0) and a negative function (either due to a negative integrand or a negative integration direction), the output values of GetUpperLimit were misleading. Although it is not possible to use GetUpperLimit for a negative function, a std::abs in IntegrateWithLog caused the return of a valid-looking value. The new treatment is that a border value (e.g. the start value of the integration or lower_lim) is returned, indicating that the actual result is out of range.