Unitizing units in all Geometries to cm. Add default argument for inner_radius to Sphere and Cylinder.
Prior to this commit, the input of json files and constructors of Geometry classes was expected to be in meter.
To avoid confusions (especially with the internal calculations of PROPOSAL as well as the output), this has been changed.
Now, all inputs are expected to be in centimeter.

This may break old configuration files that are passing distances and positions in meter.

The order of the arguments of Cylinder has been changed to allow inner_radius to be a default argument.
This may produce unnoticed errors in old scripts, therefore a log_info has been added.