Fix error missing substitution for rFlag case in EPairProduction
When calculating dEdx for EPairProduction, CalculatedEdx may run into a specific case called 'rflag' where the integral for dEdx is split in two halfs.
For the second half, the substitution v = 1-v is used for numerical reasons. However, it seems that this substitution was missing for FunctionToDEdxIntegralReverse.
This caused wrong calculations of dEdx in very specific cases, for example when calculating only-continuous dEdx values for electrons for high energies.

This breaks some test files.