Change initialization of one-dimensional dNdx interpolant.
Beforehand, the dNdx_1d interpolant has been initialized with the CrossSectionIntegral values.
However, this can cause inconsistencies when sampling stochastic losses (since the rates passed to the CalculateStochasticLoss function have been calculated using the 1d interpolant, but inside the function, these rates are compared to the values from the 2d interpolant).

This fix will make the calculation of the dNdx rates a bit more inaccurate, but consistent when sampling stochastic losses.