Use boost filesystem for backwards compability. Now the test should be buildable even on gcc7.
Delete conanfile.py and substude it through multiple conanfiles providing different usecases.
If the user want to use an not released version of proposal, he should install it lokal and use a find_package call to avoid ambiguity if the version number has not been updated and faster compile time.