Avoid unnecessary evaluations and possible floating point exceptions in Moliere scattering
For small continuous steps, xn can become negative.
This causes xn to become nan (because log(xn) is calculated), but the for loop still continues.
The case is still handled later in the code with the line 'xn != xn' (which checks if xn became nan),
however this possibly causes several unnecessary iterations of the for-loop.
Furthermore, std::log(xn) causes a floating point exception if they are enabled (FE_INVALID).
This fix catches the case where xn would become nan before the log of it is actually calculated.
Since xn will always stay nan after it once became nan, the behaviour of the code does not change
(except for avoiding the unnecessary calculations and the floating point exceptions).