Add fallback if cubic_splines::find_parameter in CrossSectionDNDXInterpolant::GetUpperLimit fails
This is very similar to the PR #185.
Under specific conditions, the NewtonRaphson method of cubic_splines::find_parameter may fail is the
spline is overshooting as the method runs into local minima.
Similarly to the fix used in PR #185, we use a bisection method as a fallback. Again, this should
only happen in very rare cases (therefore, a warning is logged).

This is also the second out of two instances where cubic_splines::find_parameter is used, so we shouldn't
run into this problem anymore.