Add comment explaining why we have to update the current_sector after a continuous step has been performed [ci skip]
If we start the propagation on a sector border facing sector X, multiple scattering might change the initial
particle direction in such a way that we don't face sector X anymore, but face and therefore enter a different sector Y instead.
This means that current_sector might change in AdvanceParticle and we need to update our references here.